docs(tailwind): move TW docs to /docs dir & add Getting Started, Config & Working With sections
