+++ docs/Tailwind/ClassnameReferences/BorderColor.stories.tsx
@@ -3,7 +3,7 @@ import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
 import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
-import { flattenEntries } from "./helpers/flatten-entries"
+import { flattenEntries } from "../helpers/flatten-entries"
 import styles from "./styles.module.scss"
 
 const prefix = "border-"
@@ -13,7 +13,7 @@ const classEntries = flattenEntries(
 )
 
 export default {
-  title: "Tailwind/Borders/Border Color",
+  title: "Tailwind/Classname References/Borders/Border Color",
   parameters: {
     docs: {
       description: {
