+++ docs/Tailwind/ClassnameReferences/TextColor.stories.tsx
@@ -3,13 +3,13 @@ import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
 import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
-import { flattenEntries } from "./helpers/flatten-entries"
+import { flattenEntries } from "../helpers/flatten-entries"
 import styles from "./styles.module.scss"
 
 const prefix = "text-"
 const classEntries = flattenEntries(prefix, kaizenTailwindTheme?.colors || {})
 export default {
-  title: "Tailwind/Typography/Text Color",
+  title: "Tailwind/Classname References/Typography/Text Color",
   parameters: {
     docs: {
       description: {
