+++ docs/Tailwind/Configuration.stories.mdx
@@ -2,7 +2,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import { CATEGORIES } from "../../storybook/constants"
 import AnatomyDiagram from './tailwind-anatomy.jpg';
 
-<Meta title={`${CATEGORIES.tailwind}/Docs/CONFIGURATION`} />
+<Meta title={`${CATEGORIES.tailwind}/Configuration`} />
 
 # Configuration
