+++ docs/Tailwind/tailwind-anatomy.jpg
