+++ storybook/tailwind.scss
@@ -1,1 +1,3 @@
 @tailwind utilities;
+@tailwind components;
+@tailwind utilities;
